Move convert files
